drm/msm/sde: Skip heavy autorefresh checks when it's not enabled
These heavy checks for seeing if autorefresh is enabled are unneeded
when the autorefresh config is disabled. These checks are performed on
every display commit and show up as using a significant amount of CPU
time in perf top. Skip them when it's unnecessary in order to improve
display rendering performance.

Change-Id: I62c12d7666409cd70aea0b959b7966cac04cb18f
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>